tests/submission/PhilipTaylor/canvas/2d.coordinatespace.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.coordinatespace</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.coordinatespace</h1>
<p class="desc">Coordinate space goes from top-left to bottom-right</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.coordinatespace">2d.coordinatespace</a>
</li></ul>
</div>
<p class="notes">This should not be upside down.
</p><p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="2d.coordinatespace.png">
</p><ul id="d"></ul>
<script>
var t = async_test("Coordinate space goes from top-left to bottom-right");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#00f';
ctx.fillRect(0, 0, 100, 50);
ctx.fillStyle = '#0ff';
ctx.fillRect(0, 0, 50, 25);
_assertPixel(canvas, 25,12, 0,255,255,255, "25,12", "0,255,255,255");
_assertPixel(canvas, 75,12, 0,0,255,255, "75,12", "0,0,255,255");
_assertPixel(canvas, 25,37, 0,0,255,255, "25,37", "0,0,255,255");
_assertPixel(canvas, 75,37, 0,0,255,255, "75,37", "0,0,255,255");
_requireManualCheck(); // because we can't tell that getPixelData isn't using the wrong coordinate space too


});
</script>

<div id="log"></div>
</body></html>
